refactor/#43 : Pick 주석 수정 & service 레이어 분리
